refactoring: moved colorspace related functions in a new file ColorSpace.cpp
