+++ src/duke/engine/ColorSpace.hpp
@@ -7,4 +7,13 @@ enum class ColorSpace
 		Auto, Linear, sRGB, GammaCorrected, AdobeRGB, Rec709, KodakLog, AlexaLogC, _END
 };
 
+// Deduce a the colorspace
+ColorSpace resolveFromExtension(const char* pFileExtension);
+ColorSpace resolveFromName(const char* pColorspace);
+
+
+// GlSl functions names
+const char* getToLinearFunction(const ColorSpace fromColorspace);
+const char* getToScreenFunction(const ColorSpace fromColorspace);
+
 } /* namespace duke */
