+++ src/duke/engine/rendering/ShaderFactory.cpp
@@ -8,6 +8,10 @@ using namespace std;
 
 namespace duke {
 
+// From ColorSpace.cpp, all the conversion functions
+// in a char *
+extern const char *pColorSpaceConversions;
+
 namespace  {
 
 std::tuple<bool, bool, bool, bool, bool, bool, ColorSpace> asTuple(const ShaderDescription &sd) {
@@ -45,25 +49,6 @@ ShaderDescription ShaderDescription::createTextureDesc(bool grayscale, bool swap
 }
 
 namespace {
-
-const char pColorSpaceConversions[] =
-		R"(
-vec3 lintolin(vec3 sample) {
-	return sample;
-}
-vec3 alexatolin(vec3 sample) {
-    return mix(pow(vec3(10.0),((sample-0.385537)/0.247190)-0.052272)/5.555556 , (sample-0.092809)/5.367655, lessThan(sample, vec3(0.149658)));
-}
-vec3 cineontolin(vec3 sample) {
-	return 1.010915615730753*(pow(vec3(10), (1023*sample-685)/300)-0.010797751623277);
-}
-vec3 srgbtolin(vec3 sample) {
-    return mix(pow((sample+0.055)/1.055,vec3(2.4)), sample/12.92, lessThan(sample, vec3(0.04045)));
-}
-vec3 lintosrgb(vec3 sample) {
-    sample = mix((1.055*pow(sample,vec3(1/2.4)))-vec3(0.055), 12.92*sample, lessThan(sample, vec3(0.0031308)));
-    return clamp(sample, vec3(0), vec3(1));
-})";
 
 const char pSampleTenbitsUnpack[] =
 		R"(
@@ -192,36 +177,6 @@ void main(void)
 }
 )";
 
-const char* getToLinearFunction(const ColorSpace fromColorspace) {
-    switch (fromColorspace) {
-        case ColorSpace::AlexaLogC:
-            return "alexatolin";
-        case ColorSpace::KodakLog:
-            return "cineontolin";
-        case ColorSpace::Linear:
-            return "lintolin";
-        case ColorSpace::sRGB:
-        case ColorSpace::GammaCorrected:
-            return "srgbtolin";
-        case ColorSpace::Auto:
-        default:
-            throw std::runtime_error("ColorSpace must be resolved at this point");
-    }
-}
-
-const char* getToScreenFunction(const ColorSpace fromColorspace) {
-    switch (fromColorspace) {
-        case ColorSpace::AlexaLogC:
-        case ColorSpace::KodakLog:
-        case ColorSpace::Linear:
-        case ColorSpace::sRGB:
-        case ColorSpace::GammaCorrected:
-            return "lintosrgb";
-        case ColorSpace::Auto:
-        default:
-            throw std::runtime_error("ColorSpace must be resolved at this point");
-    }
-}
 
 void appendToLinearFunction(ostream&stream, const ColorSpace colorspace) {
 	stream << endl << "vec3 toLinear(vec3 sample){return " << getToLinearFunction(colorspace) << "(sample);}" << endl;
